+++ comercio/src/main/java/com/feguzman/comercio/Entity/DetailOfSaleForLess.java
@@ -22,8 +22,8 @@ public class DetailOfSaleForLess {
     private int unitQuantity;
 
     @Column(name = "cost")
-    private int cost;
+    private double cost;
 
     @Column(name = "totalCost")
-    private int totalCost;
+    private double totalCost;
 }
